signature overlay coordinates, bug where the leftwould show less signatures
